remove seed peers, refactor
