+++ core/src/config.rs
@@ -43,104 +43,7 @@ const fn parse_version() -> [u8; 3] {
     out
 }
 
-pub const SEED_NODES: &[&str] = &[
-    "104.218.45.23",
-    "136.243.51.155",
-    "139.60.162.57",
-    "152.114.194.143",
-    "152.114.194.161",
-    "152.114.194.195",
-    "156.67.62.246",
-    "176.9.24.34",
-    "178.237.58.137",
-    "178.237.58.139",
-    "178.237.58.203",
-    "178.237.58.217",
-    "178.237.58.218",
-    "185.105.111.184",
-    "185.130.227.22",
-    "185.148.38.16",
-    "188.213.129.120",
-    "194.147.214.237",
-    "194.147.214.238",
-    "194.164.217.150",
-    "194.180.188.146",
-    "194.180.207.108",
-    "194.180.207.39",
-    "195.26.231.137",
-    "195.26.231.9",
-    "205.172.57.173",
-    "205.172.59.6",
-    "212.108.83.119",
-    "212.108.83.192",
-    "212.108.83.245",
-    "212.237.219.142",
-    "212.237.219.67",
-    "217.123.171.82",
-    "31.207.47.135",
-    "35.91.34.26",
-    "35.94.234.66",
-    "37.228.93.218",
-    "37.27.238.10",
-    "37.27.238.30",
-    "38.83.202.35",
-    "38.83.202.36",
-    "38.83.202.37",
-    "38.83.202.38",
-    "38.83.202.39",
-    "38.83.202.40",
-    "38.83.202.41",
-    "38.83.202.42",
-    "38.83.202.43",
-    "38.83.202.44",
-    "38.83.202.45",
-    "38.83.202.46",
-    "38.83.202.47",
-    "38.83.202.48",
-    "38.83.202.49",
-    "38.83.202.53",
-    "38.83.202.62",
-    "38.83.202.63",
-    "38.83.202.64",
-    "38.83.202.66",
-    "46.17.103.22",
-    "46.17.96.23",
-    "46.4.13.251",
-    "5.39.221.243",
-    "62.109.16.125",
-    "62.169.17.132",
-    "66.151.40.64",
-    "66.248.204.226",
-    "67.222.138.149",
-    "67.222.138.150",
-    "67.222.157.66",
-    "67.222.157.68",
-    "72.9.144.110",
-    "72.9.146.100",
-    "72.9.146.101",
-    "72.9.146.102",
-    "72.9.146.103",
-    "72.9.146.104",
-    "72.9.146.105",
-    "72.9.146.106",
-    "72.9.146.107",
-    "72.9.146.108",
-    "72.9.146.109",
-    "72.9.146.110",
-    "72.9.146.111",
-    "72.9.146.112",
-    "72.9.146.113",
-    "72.9.146.114",
-    "72.9.146.115",
-    "72.9.146.116",
-    "72.9.146.70",
-    "72.9.146.71",
-    "72.9.146.72",
-    "72.9.146.73",
-    "72.9.146.75",
-    "72.9.146.76",
-    "72.9.146.77",
-];
+pub const SEED_NODES: &[&str] = &["104.218.45.23", "72.9.144.110"];
 
 // seed anr from elixir config/config.exs
 #[derive(Debug, Clone, Serialize, Deserialize)]
