+++ core/src/lib.rs
@@ -1,7 +1,5 @@
 #![allow(dead_code)]
 
-use crate::node::{anr, peers};
-
 pub mod bic;
 pub mod config;
 pub mod consensus;
@@ -13,23 +11,6 @@ pub mod socket;
 pub mod utils;
 pub mod wasm;
 
+pub use config::Config;
 pub use context::{Context, PeerInfo};
 pub use metrics::{MetricsSnapshot, UdpStats};
-
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    #[error(transparent)]
-    Fabric(#[from] consensus::fabric::Error),
-    #[error(transparent)]
-    Archiver(#[from] utils::archiver::Error),
-    #[error(transparent)]
-    Protocol(#[from] node::protocol::Error),
-    #[error(transparent)]
-    Config(#[from] config::Error),
-    #[error(transparent)]
-    Anr(#[from] anr::Error),
-    #[error(transparent)]
-    Peers(#[from] peers::Error),
-    #[error("{0}")]
-    String(String),
-}
